+++ ImdbBackend/DataLayer/Ratings/Rating.cs
@@ -10,7 +10,7 @@ public class Rating
     [ForeignKey("Title")]
     public string TConst { get; set; } = string.Empty;
 
-    public double AverageRating { get; set; }
+    public decimal AverageRating { get; set; }
 
     public int NumberOfVotes { get; set; }
 
@@ -24,7 +24,7 @@ public class RatingForUserResult
 
     public DateTime TimeStamp { get; set; }
 
-    public double Rating { get; set; }
+    public decimal Rating { get; set; }
 
 }
 
@@ -36,5 +36,5 @@ public class AddRatingResult
 
 public class PersonRatingResult
 {
-    public double PersonRating { get; set; }
+    public decimal PersonRating { get; set; }
 }
